requirements.txt
amqp==5.0.2
billiard==3.6.3.0
celery==5.0.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
dj-database-url==0.5.0
Django==2.2
django-filter==2.4.0
django-instagram==0.3.2
django-model-search==2.2
django-modeltranslation==0.15.2
# django-novaposhta==2.3
django-smart-selects==1.5.9
django-widget-tweaks==1.4.8
gunicorn==20.0.4
idna==2.10
kombu==5.0.2
lxml==4.6.1
Pillow==7.2.0
postgres==3.0.0
prompt-toolkit==3.0.8
psycopg2==2.8.6
psycopg2-binary==2.8.6
psycopg2-pool==1.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
